MediBot-main/server/backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
mysql-connector-python==9.0.0
bcrypt==4.2.0
python-dotenv==1.0.1
google-auth==2.37.0
google-auth-oauthlib==1.2.1

